flask/requirements.txt
Flask==1.0.2
flask_cors==3.0.8
fpdf==1.7.2
pdfkit==0.6.1
PyPDF2==1.26.0
#wkhtmltopdf==0.2
python-owasp-zap-v2.4==0.0.16
python-nmap==0.6.1
rq==1.3.0
redis==3.4.1
Werkzeug==0.14.1
pdfrw==0.4
APScheduler==3.6.3
wapiti3==3.0.3
selenium==3.141.0 
PyVirtualDisplay==0.2.5
rq-dashboard==0.6.1